requirements.txt
requests
pydub
fastapi
torch
torchvision
torchaudio
python-multipart
uvicorn
swagger-ui-py
